A nasty hack for story 0
Unfortunately there was a screw up and we commited awhile back from story_0
when we shouldn't have and then made things worse by doing a revert. Because
of the revert merging with the story_0 branch is a plain nightmare. Since
we plan on getting rid of all these files anyway rather than doing a forced
reset on the HEAD I'm just copying over the story_0 files manually (so we
lose all the versioning history on the master branch). Eventually we are
going to nuke the entire depot since it contains a bunch of files we
don't need and don't want to make people downloaded.

So that's why I'm o.k. with this hack for now.