Make it easier to run on localhost.
The main fix is to assign the mock object to `global.Mobile` rather than
to defining it with `var Mobile`, because otherwise, the definition would apply
only within the current module.

The trigger for the mockups was moved behind a more explicit MOCK_MOBILE
environment variable rather than NODE_ENV, because the latter is a fairly
commonly used and we don't want these mocks to accidentally apply, for example,
in a Cordova-based environment.

The readme is updated with steps needed after a fresh clone to be able to run
on localhost.

Fixes #58.